move osmChange state to separate redux module
