+++ app/assets/scripts/actions/action-creators.js
@@ -1,11 +1,7 @@
 import fetch from 'isomorphic-fetch';
 import _ from 'lodash';
-import XmlReader from 'xml-reader';
 import * as actions from './action-types';
 import config from '../config';
-import {
-  fetchWayTaskEpic
-} from '../redux/modules/tasks';
 
 
 // ////////////////////////////////////////////////////////////////
@@ -169,101 +165,6 @@ export function fetchAdminStats (id = null) {
   };
 }
 
-
-// ////////////////////////////////////////////////////////////////
-//                        osm changesets                         //
-// ////////////////////////////////////////////////////////////////
-const requestOsmChange = () => ({ type: actions.REQUEST_OSM_CHANGE });
-
-const createChangeset = (dispatch, cb) => {
-  const changesetUrl = `${config.api}/changeset/create`;
-  const details = {
-    uid: 555555,
-    user: 'Openroads Tasks'
-  };
-  return fetch(changesetUrl, {
-    method: 'PUT',
-    body: objectToBlob(details)
-  }).then(response => {
-    if (response.status >= 400) {
-      throw new Error('Bad response');
-    }
-    return response.text();
-  }).then(cb, e => {
-    return dispatch(completeOsmChange(null, e));
-  });
-};
-
-const uploadChangeset = (dispatch, taskId, payload, changesetId) => {
-  return fetch(`${config.api}/changeset/${changesetId}/upload`, {
-    method: 'POST',
-    body: objectToBlob({ osmChange: payload })
-  })
-  .then(() => {
-    // TODO - completeOsmChange shouldn't run until after fetchWayTask refreshes task geometry
-    dispatch(fetchWayTaskEpic(taskId));
-    dispatch(completeOsmChange(taskId));
-  });
-};
-
-const completeOsmChange = (taskId, error = null) => ({
-  type: actions.COMPLETE_OSM_CHANGE,
-  taskId,
-  error,
-  receivedAt: Date.now()
-});
-
-const objectToBlob = (obj) => {
-  return new Blob([JSON.stringify(obj, null, 2)], {type: 'application/json'});
-};
-
-export function queryOsm (taskId, payload) {
-  return function (dispatch) {
-    dispatch(requestOsmChange());
-    createChangeset(dispatch, changesetId => uploadChangeset(dispatch, taskId, payload, changesetId));
-  };
-}
-
-// Since the geojson that powers the tasks endpoint doesn't include node IDs,
-// we must query the XML endpoint to get these IDs, then format them into
-// a `delete` action.
-export function deleteEntireWays (taskId, wayIds) {
-  return function (dispatch) {
-    dispatch(requestOsmChange());
-
-    const nodeIds = [];
-    fetch(`${config.api}/api/0.6/ways?nodes=true&excludeDoubleLinkedNodes=true&ways=${wayIds.join(',')}`)
-    .then(response => {
-      if (response.status >= 400) {
-        throw new Error('Bad response');
-      }
-      return response.text();
-    }).then(parseXml, e => {
-      return dispatch(completeOsmChange(null, e));
-    });
-
-    function parseXml (xmlDoc) {
-      let reader = XmlReader.create({stream: true});
-      reader.on('tag:node', node => {
-        if (node.attributes.id) {
-          nodeIds.push(node.attributes.id);
-        }
-      });
-      reader.on('done', formatPayload);
-      reader.parse(xmlDoc);
-    }
-
-    function formatPayload () {
-      let payload = {
-        delete: {
-          node: nodeIds.map(id => ({ id })),
-          way: wayIds.map(id => ({ id }))
-        }
-      };
-      createChangeset(dispatch, changesetId => uploadChangeset(dispatch, taskId, payload, changesetId));
-    }
-  };
-}
 
 // ////////////////////////////////////////////////////////////////
 //                        project tasks                          //
