+++ app/assets/scripts/actions/action-types.js
@@ -23,9 +23,6 @@ export const SET_FILTERED_VPROMMS = 'SET_FILTERED_VPROMMS';
 export const REQUEST_ADMIN_STATS = 'REQUEST_ADMIN_STATS';
 export const RECEIVE_ADMIN_STATS = 'RECEIVE_ADMIN_STATS';
 
-export const REQUEST_OSM_CHANGE = 'REQUEST_OSM_CHANGE';
-export const COMPLETE_OSM_CHANGE = 'COMPLETE_OSM_CHANGE';
-
 export const REQUEST_PROJECTS = 'REQUEST_PROJECTS';
 export const RECEIVE_PROJECTS = 'RECEIVE_PROJECTS';
