+++ app/assets/scripts/redux/reducer.js
@@ -2,9 +2,8 @@ import { combineReducers } from 'redux';
 import { routeReducer } from 'react-router-redux';
 import _ from 'lodash';
 import * as actions from '../actions/action-types';
-import waytasks, {
-  RELOAD_WAY_TASK
-} from './modules/tasks';
+import waytasks from './modules/tasks';
+import osmChange from './modules/osm';
 import { ADMIN_MAP } from '../constants';
 
 
@@ -105,41 +104,6 @@ const stats = function (state = {fetching: false, fetched: false, data: null}, a
   return state;
 };
 
-
-const osmChangeState = {
-  fetching: false,
-  fetched: false,
-  taskId: null,
-  error: null
-};
-const osmChange = function (state = osmChangeState, action) {
-  switch (action.type) {
-    case actions.REQUEST_OSM_CHANGE:
-      console.log('REQUEST_OSM_CHANGE');
-      state = _.cloneDeep(state);
-      state.error = null;
-      state.fetching = true;
-      break;
-    case actions.COMPLETE_OSM_CHANGE:
-      console.log('COMPLETE_OSM_CHANGE');
-      state = _.cloneDeep(state);
-      if (action.error) {
-        state.error = action.error;
-      } else {
-        state.taskId = action.taskId;
-      }
-      state.fetching = false;
-      state.fetched = true;
-      break;
-    case RELOAD_WAY_TASK:
-      console.log('RELOAD_WAY_TASK');
-      state = _.cloneDeep(state);
-      state.fetched = false;
-      state.taskId = false;
-      break;
-  }
-  return state;
-};
 
 const projecttasksDefaultState = {
   fetching: false,
